How to redirect to html page after login with facebook
good afternoon,

I have been trying to use “login with facebook” on my website for 15 days. I have been using the code published by Facebook
My problem with this code is that I want the customer after login with his facebook (through the window that pop out) to be redirected to a specific page on my website, but what happens is that after the customer does login with your facebook, the client is redirected to the same page that asks him to login.

I have already followed the steps that are explained in the link that I mentioned earlier. Could someone help me? Thanks

The code i am using is this: